Move datacall to a file, datacall.go
